+++ src/main/typescript/lang.d.ts
@@ -604,6 +604,8 @@ export interface LangSchema {
             missing: string;
             /** An error message for an unknown error. */
             unknown: string;
+            /** An error message for an unknown RuleElement condition. */
+            unknownCondition: string;
             /** An error message for an unknown RuleElement hook. */
             unknownHook: string;
             /** An error message for an unknown RuleElement type. */